third_party/blink/web_tests/fast/forms/form-dirname-attribute-expected.txt

"dirname" attribute

PASS input.dirName is "Hello"
PASS textArea.dirName is "Hello"
PASS successfullyParsed is true

TEST COMPLETE
